Fix compilation errors on 64bit machines.

+++ src/examples/list_switches/list_switches.c
@@ -36,7 +36,7 @@ recv_reply( uint16_t tag, void *data, size_t len, void *user_data ) {
   info( "[%s]", time_string );
 
   uint64_t *dpid = ( uint64_t * ) data;
-  unsigned int count = len / sizeof( uint64_t );
+  size_t count = len / sizeof( uint64_t );
   unsigned int i;
   for ( i = 0; i < count; i++ ) {
     info( "%#" PRIx64, ntohll( dpid[ i ] ) );
